fixed some step 1 errors
- numbers are now treaded as int
- exception for end of input when parenthese is open

+++ impls/cpp.2/reader.hpp
@@ -16,7 +16,7 @@ class Reader {
 private:
 
     std::vector<std::string> tokens;
-    int cur_pos;
+    long unsigned int cur_pos;
 
     std::string next(void);
